Change variable to iterate over coils to 16-bit :bangbang:

+++ src/slave/scoils.c
@@ -31,7 +31,7 @@ uint8_t modbusParseRequest01( ModbusSlave *status, union ModbusParser *parser )
 	//Update frame length
 	uint8_t frameLength = 8;
 	uint8_t coil = 0;
-	uint8_t i = 0;
+	uint16_t i = 0;
 
 	//Check if given pointers are valid
 	if ( status == NULL ) return MODBUS_ERROR_OTHER;
@@ -217,7 +217,7 @@ uint8_t modbusParseRequest15( ModbusSlave *status, union ModbusParser *parser )
 	//Using data from union pointer
 
 	//Update frame length
-	uint8_t i = 0;
+	uint16_t i = 0;
 	uint8_t frameLength;
 	uint8_t coil = 0;
